+++ test/e2e/specs/site-editor/global-styles-sidebar.spec.js
@@ -18,14 +18,14 @@ test.describe( 'Global styles sidebar', () => {
 	} );
 
 	test( 'should filter blocks list results', async ( { page } ) => {
-		// Navigate to Styles -> Blocks.
+		// Navigate to Styles -> Blocks & Elements.
 		await page
 			.getByRole( 'region', { name: 'Editor top bar' } )
 			.getByRole( 'button', { name: 'Styles' } )
 			.click();
 		await page
 			.getByRole( 'region', { name: 'Editor settings' } )
-			.getByRole( 'button', { name: 'Blocks' } )
+			.getByRole( 'button', { name: 'Blocks & Elements' } )
 			.click();
 
 		await page